fix: Verify and fix misconceptions for students 21-30 (200120-200129)
Fixed mismatches:
- Lee_Luke_200120 Q2: Added OTHER002 (hardcoded values)
- Lee_Luke_200120 Q3: Added DT003 (nextInt for doubles)
- Nelson_Avery_200123 Q2: Added OTHER002 (hardcoded values)
- Nelson_Avery_200123 Q3: Added INPUT001 (commented out import)
- Wright_Evelyn_200125 Q1: Added INPUT003 (not reading all inputs)
- Wright_Evelyn_200125 Q2: Added VAR001 (wrong formula)

+++ authentic_seeded/Baker_Oliver_200129_DT003/Q2.java
@@ -3,15 +3,16 @@
 // Assignment 2
 
 import java.util.Scanner;
+
 public class Q2 {
    public static void main(String[] args) {
       Scanner scanner = new Scanner(System.in);
       System.out.print("Enter the driving distance in miles: ");
-      double distance = scanner.nextDouble();
+      double distance = scanner.nextInt();
       System.out.print("Enter miles per gallon: ");
-      double milesPerGallon = scanner.nextDouble();
+      double milesPerGallon = scanner.nextInt();
       System.out.print("Enter price in $ per gallon: ");
-      double price = scanner.nextDouble();
+      double price = scanner.nextInt();
       double cost = (distance / milesPerGallon) * price;
       System.out.println("The cost of driving is $" + cost);
    }
