+++ authentic_seeded/Carter_Olivia_200128_DT003/Q1.java
@@ -8,9 +8,9 @@ public class Q1 {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		System.out.print("Enter v0, v1, and t: ");
-		double vStart = scan.nextDouble();
-		double vEnd = scan.nextDouble();
-		double t = scan.nextDouble();
+		double vStart = scan.nextInt();
+		double vEnd = scan.nextInt();
+		double t = scan.nextInt();
 		double a = (vEnd - vStart) / t;
 		System.out.println("The average acceleration is " + a);
 	}
